增加进度数字的显示

+++ app/src/main/java/com/blackflagbin/myapplication/MainActivity.kt
@@ -9,10 +9,9 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
         bt_start.setOnClickListener {
             if (et_progress.text.toString().isNotEmpty()) {
-                progress.updateProgress(et_progress.text.toString().toFloat())
+                progress.updateProgress(et_progress.text.toString().toInt())
             }
         }
     }
